詳細の画面遷移グラフ変更
